+++ include/ldt/point_2d.h
@@ -62,6 +62,16 @@ struct point_2d
 	}
 
 	//!Gets a new point by substraction of components.
+	point_2d operator+(T _p) const {
+		return point_2d<T> {x+_p, y+_p};
+	}
+
+	point_2d& operator+=(T _p) {
+		x+=_p;
+		y+=_p;
+		return *this;
+	}
+
 	point_2d operator-(T p) const {
 		return point_2d<T> {x-p, y-p};
 	}
@@ -78,16 +88,15 @@ struct point_2d
 		return point_2d<T> {x*p, y*p};
 	}
 
+	point_2d& operator*=(T _p) { 
+		x*=_p;
+		y*=_p;
+		return *this;
+	}
+
 	point_2d operator*(const point_2d& p) const {
 
 		return point_2d<T> {x*p.x, y*p.y};
-	}
-
-	//!Multiplies the point.
-	point_2d& operator*=(T p) { 
-		x*=p;
-		y*=p;
-		return *this;
 	}
 
 	point_2d& operator*=(const point_2d& p) {